requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
prometheus-client==0.17.1
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
faker==20.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
pre-commit==3.5.0

# Monitoring & Logging
sentry-sdk==1.32.0
python-json-logger==2.0.7

# AI/ML Dependencies
openai==1.3.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.1.1

# Email & Communication
smtplib3==0.1.0
twilio==8.10.0
linkedin-api==2.0.0a5

# Web Scraping & Data Collection
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
playwright==1.39.0

# Social Media Integration
tweepy==4.14.0
instagram-private-api==1.6.0
python-linkedin-v2==0.9.4 